+++ crates/ethereum/cli/Cargo.toml
@@ -8,3 +8,16 @@ homepage.workspace = true
 repository.workspace = true
 
 [lints]
+workspace = true
+
+[dependencies]
+reth-cli.workspace = true
+reth-chainspec.workspace = true
+
+alloy-genesis.workspace = true
+
+eyre.workspace = true
+
+shellexpand.workspace = true
+serde_json.workspace = true
+clap = { workspace = true, features = ["derive", "env"] }
