+++ crates/ethereum/cli/src/lib.rs
@@ -7,3 +7,6 @@
 )]
 #![cfg_attr(not(test), warn(unused_crate_dependencies))]
 #![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
+
+/// Chain specification parser.
+pub mod chainspec;
