+++ crates/optimism/cli/Cargo.toml
@@ -12,7 +12,6 @@ workspace = true
 
 [dependencies]
 reth-static-file-types = { workspace = true, features = ["clap"] }
-clap = { workspace = true, features = ["derive", "env"] }
 reth-cli-commands.workspace = true
 reth-consensus.workspace = true
 reth-db = { workspace = true, features = ["mdbx"] }
@@ -26,18 +25,25 @@ reth-static-file.workspace = true
 reth-execution-types.workspace = true
 reth-node-core.workspace = true
 reth-primitives.workspace = true
-
-
 reth-stages-types.workspace = true
 reth-node-events.workspace = true
 reth-network-p2p.workspace = true
 reth-errors.workspace = true
-
 reth-config.workspace = true
-alloy-primitives.workspace = true
-futures-util.workspace = true
 reth-evm-optimism.workspace = true
+reth-cli.workspace = true
+reth-chainspec.workspace = true
 
+# eth
+alloy-genesis.workspace = true
+alloy-primitives.workspace = true
+
+
+# misc
+shellexpand.workspace = true
+serde_json.workspace = true
+futures-util.workspace = true
+clap = { workspace = true, features = ["derive", "env"] }
 
 
 tokio = { workspace = true, features = [
